Disable temporarily the UIAutomator submodule
The current configuration of UIAutomator defined in gradle doesn't work out of
the box as it relies in the old sdk.dir property to find the Android SDK
directory.

The fix is easy, but probably it would be better to just migrate to Espresso, so
for now we'll disable this submodule to adapt the rest of the app first.